REFACTOR: 패키지 정리
See also: #1
