+++ src/main/java/com/sparta/deliveryi/global/exception/GlobalExceptionHandler.java
@@ -1,5 +1,6 @@
-package com.sparta.deliveryi.global;
+package com.sparta.deliveryi.global.exception;
 
+import com.sparta.deliveryi.global.presentation.dto.ApiResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
