+++ src/main/java/com/sparta/deliveryi/global/infrastructure/MessageResolver.java
@@ -1,4 +1,4 @@
-package com.sparta.deliveryi.global;
+package com.sparta.deliveryi.global.infrastructure;
 
 import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Autowired;
